Finished Main Homework and Bonus (Questions) - hopefully with the release of spark 3.4 I will be able to commit the Bonus-Coding part as well.
Although the Code works, I find it not very satisfying that in the scenario of changing the schema of the session several parts of the code would have to be adjusted. This seems to be something that is just prone to mistakes. Do you have an idea on how to improve the code to make it a bit more resilient?

Cheers,
Colin